# swagger.yaml
swagger: '2.0'
info:
  description: API simplificada baseada na API Consents do Open Banking
  version: 1.0.0
  title: Consents API
  contact:
    name: Micheline Barros
    email: micheline.barros@sensedia.com
  license:
    name: Apache License Version 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
host: localhost:8080
basePath: /
tags:
  - name: consents-controller
    description: Consents Controller
paths:
  /consents/v1/consents:
    post:
      tags:
        - consents-controller
      summary: createConsent
      operationId: createConsentUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: consents
          description: consents
          required: true
          schema:
            $ref: '#/definitions/Consents'
      responses:
        '201':
          description: Consentimento criado com sucesso.
          schema:
            $ref: '#/definitions/ConsentsDTO'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: O recurso solicitado não existe ou não foi implementado
          schema:
            $ref: '#/definitions/ResponseErrorDetails'
  /consents/v1/consents/{consentId}:
    get:
      tags:
        - consents-controller
      summary: consentsDetails
      operationId: consentsDetailsUsingGET
      produces:
        - '*/*'
      parameters:
        - name: consentId
          in: path
          description: consentId
          required: true
          type: string
      responses:
        '200':
          description: Consentimento consultado com sucesso.
          schema:
            $ref: '#/definitions/Consents'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: O recurso solicitado não existe ou não foi implementado
          schema:
            $ref: '#/definitions/ResponseErrorDetails'
    delete:
      tags:
        - consents-controller
      summary: deleteConsents
      operationId: deleteConsentsUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: consentId
          in: path
          description: consentId
          required: true
          type: string
      responses:
        '204':
          description: Consentimento revogado com sucesso.
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Consentimento revogado com sucesso.
          schema:
            $ref: '#/definitions/ResponseErrorDetails'
definitions:
  Consents:
    type: object
    required:
      - businessEntity
      - loggedUser
      - transactionToDateTime
    properties:
      businessEntity:
        type: string
        pattern: ([0-9]{2}[.]?[0-9]{3}[.]?[0-9]{3}[/]?[0-9]{4}[-]?[0-9]{2})
      id:
        type: string
      loggedUser:
        type: string
      transactionFromDateTime:
        type: string
        format: date
      transactionToDateTime:
        type: string
        format: date
    title: Consents
  ConsentsDTO:
    type: object
    properties:
      id:
        type: string
      transactionFromDateTime:
        type: string
        format: date
      transactionToDateTime:
        type: string
        format: date
    title: ConsentsDTO
  ErrorDetail:
    type: object
    properties:
      code:
        type: string
      detail:
        type: string
      title:
        type: string
    title: ErrorDetail
  ErrorMeta:
    type: object
    properties:
      requestDateTime:
        type: string
        format: date-time
      totalPage:
        type: integer
        format: int32
      totalRecords:
        type: integer
        format: int32
    title: ErrorMeta
  ResponseErrorDetails:
    type: object
    properties:
      errors:
        type: array
        items:
          $ref: '#/definitions/ErrorDetail'
      meta:
        $ref: '#/definitions/ErrorMeta'
    title: ResponseErrorDetails